backend/api/routes/workflows.py
"""Workflow management endpoints"""

from fastapi import APIRouter, HTTPException
from typing import Optional
import yaml
import uuid
from datetime import datetime

router = APIRouter()

# In-memory workflow storage
WORKFLOWS_DB = {}

BUILTIN_WORKFLOWS = {
    'r6g-detection': {
        'id': 'r6g-detection',
        'name': 'R6G Detection Pipeline',
        'description': 'Complete pipeline for R6G detection',
        'author': 'sers-team',
        'category': 'detection',
        'downloads': 1234,
        'rating': 4.8,
    },
}

@router.get("/workflows")
async def list_workflows(category: Optional[str] = None):
    """List available workflows"""
    workflows = list(BUILTIN_WORKFLOWS.values()) + list(WORKFLOWS_DB.values())
    if category and category != 'all':
        workflows = [w for w in workflows if w.get('category') == category]
    return {"workflows": workflows}

@router.get("/workflows/{workflow_id}")
async def get_workflow(workflow_id: str):
    """Get a specific workflow"""
    if workflow_id in BUILTIN_WORKFLOWS:
        return BUILTIN_WORKFLOWS[workflow_id]
    if workflow_id in WORKFLOWS_DB:
        return WORKFLOWS_DB[workflow_id]
    raise HTTPException(status_code=404, detail="Workflow not found")

@router.post("/workflows/{workflow_id}/execute")
async def execute_workflow(workflow_id: str):
    """Execute a workflow"""
    return {"success": True, "workflow_id": workflow_id, "status": "completed"}
