backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
plotly>=5.18.0
openpyxl>=3.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.1
pyyaml>=6.0.1
redis>=5.0.0
celery>=5.3.0
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.4.0
